requirements.txt
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
more-itertools==4.3.0
Pillow==5.3.0
python-docx==0.8.7
six==1.11.0
transitions==0.6.8
